requirements.txt
anyio==4.11.0
appnope==0.1.4
asttokens==3.0.0
certifi==2025.10.5
click==8.3.0
comm==0.2.1
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.1
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
ftfy==6.3.1
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.1.2
idna==3.11
ipykernel==6.30.1
ipython==8.30.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
open_clip_torch==3.2.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
pillow==12.0.0
pip==25.2
platformdirs==4.5.0
prompt_toolkit==3.0.52
protobuf==6.33.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pymilvus==2.6.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
regex==2025.11.3
safetensors==0.6.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
stack_data==0.6.3
sympy==1.14.0
timm==1.0.22
torch==2.9.0
torchvision==0.24.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
wcwidth==0.2.13
wheel==0.45.1
